+++ src/main/java/lotto/domain/Rank.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.Arrays;
 
@@ -6,8 +6,8 @@ public enum Rank {
     FIRST(6, 2_000_000_000),
     SECOND(5, 30_000_000),
     THIRD(5, 1_500_000),
-    FOURTH(4, 50000),
-    FIFTH(3, 5000);
+    FOURTH(4, 50_000),
+    FIFTH(3, 5_000);
 
     private final int matched;
     private final int prize;
