+++ src/main/java/lotto/domain/WinningNumber.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import lotto.util.StringToIntConverter;
